Upd old a--find with end boundary check
We try to read the action state (which is already
under-featured but over-complicated) and read the
`:bound-end?` attribute on it, based on the pattern
input, to determine whether we should match strictly
or loosely to the last symbol in the pattern def a
la regex `$`. The logic is added, but the overall
design is worse.

Adding just for record, but to be deprecated along
with the a--* functions