Fix bug where async requests get flooded, split :count
With only one :count field to track number of requests,
on the first request, :count is 0, resulting in no delay
in the request. But if the request is async, the go loop
will reach the :count = 0 check almost immediately, which
causes a flood of requests to get queued.

Dividing :count into :response-count (the original intended
meaning) and :request-count (the correct check for = 0)
fixes this issue.