Update swimlane example to imply a db-driven architecture
We assume the node attributes can be individually set and
queried, and picked up at render-time.

Suppose we read the graph data from a series of `.org` files.
node `id` is an arbitrary uuid, or filename-based unique string,
and "label" is a title field. The width/height/color/section and
other attributes can be set e.g. in the `#+PROPERTIES`
drawer and applied at render time.